refactor(remote-layers): rearrange remote layer, wmts, wms classes
